+++ crates/entrypoint/src/cli/command/azure/role/assignment/mod.rs
@@ -1,10 +1,10 @@
-pub mod azure_role_assignment;
-pub mod azure_role_assignment_browse;
-pub mod azure_role_assignment_list;
+pub mod azure_role_assignment_cli;
+pub mod azure_role_assignment_browse_cli;
+pub mod azure_role_assignment_list_cli;
 
-pub use azure_role_assignment::AzureRoleAssignmentCommand;
-pub use azure_role_assignment_browse::AzureRoleAssignmentBrowseArgs;
-pub use azure_role_assignment_list::AzureRoleAssignmentListArgs;
+pub use azure_role_assignment_cli::AzureRoleAssignmentCommand;
+pub use azure_role_assignment_browse_cli::AzureRoleAssignmentBrowseArgs;
+pub use azure_role_assignment_list_cli::AzureRoleAssignmentListArgs;
 use clap::Args;
 use eyre::Result;