+++ crates/entrypoint/src/cli/command/azure_devops/group/list/azure_devops_group_list_cli.rs
@@ -1,6 +1,6 @@
 use clap::Args;
 use cloud_terrastodon_azure_devops::prelude::AzureDevOpsProjectArgument;
-use cloud_terrastodon_azure_devops::prelude::fetch_azure_devops_groups;
+use cloud_terrastodon_azure_devops::prelude::fetch_azure_devops_groups_for_project;
 use cloud_terrastodon_azure_devops::prelude::get_default_organization_url;
 use eyre::Result;
 use serde_json::to_writer_pretty;
@@ -16,7 +16,7 @@ pub struct AzureDevOpsGroupListArgs {
 impl AzureDevOpsGroupListArgs {
     pub async fn invoke(self) -> Result<()> {
         let org_url = get_default_organization_url().await?;
-        let groups = fetch_azure_devops_groups(&org_url, self.project).await?;
+        let groups = fetch_azure_devops_groups_for_project(&org_url, self.project).await?;
         to_writer_pretty(stdout(), &groups)?;
         Ok(())
     }