+++ crates/entrypoint/src/cli/command/azure_devops/project/dump/mod.rs
@@ -1,0 +1,2 @@
+mod azure_devops_project_dump_cli;
+pub use azure_devops_project_dump_cli::*;