+++ crates/entrypoint/src/cli/command/azure_devops/project/mod.rs
@@ -1,4 +1,5 @@
 mod azure_devops_project_cli;
 pub mod list;
 pub mod show;
+pub mod dump;
 pub use azure_devops_project_cli::*;