+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Rename `get_azure_devops_access_token` to `fetch_azure_devops_personal_access_token`
 - Rename `get_azure_devops_pat` to `get_azure_devops_personal_access_token_from_credential_manager`
 - Add `ct az devops agent package show-newest` command
+- Rename `fetch_azure_devops_groups` to `fetch_azure_devops_groups_for_project`
+- Make role definition browse command reloadable
 
 # v0.31.0